Deduplicate radiation convolution types via public header
